+++ test/pubsub/local_pubsub_test.exs
@@ -1,0 +1,33 @@
+defmodule Commanded.PubSub.LocalPubSubTest do
+  use ExUnit.Case
+
+  alias Commanded.PubSub.LocalPubSub
+
+  @topic "test"
+
+  setup do
+    Application.put_env(:commanded, :pubsub, :local)
+
+    on_exit(fn ->
+      Application.delete_env(:commanded, :pubsub)
+    end)
+  end
+
+  describe "pub/sub" do
+    test "should receive broadcast message" do
+      assert :ok = LocalPubSub.subscribe(@topic)
+      assert :ok = LocalPubSub.broadcast(@topic, :message)
+
+      assert_receive(:message)
+    end
+  end
+
+  describe "tracker" do
+    test "should list tracked processes" do
+      self = self()
+
+      assert :ok = LocalPubSub.track(@topic, :example)
+      assert [{:example, ^self}] = LocalPubSub.list(@topic)
+    end
+  end
+end